Removido arquivos cpp de classes templates e colocados a implementação desses métodos no arquivo cabeçalho.
